pyproject.toml
[project]
name = "bio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.85",
    "matplotlib>=3.10.5",
    "pandas>=2.3.1",
    "primer3-py>=2.2.0",
]
